+++ CMakeLists.txt
@@ -38,7 +38,7 @@ pkg_check_modules(LibPCAP REQUIRED libpcap>=1.10)
 pkg_check_modules(OpenSSL REQUIRED openssl>=3.0)
 pkg_check_modules(TCLAP REQUIRED tclap>=1.2)
 
-pkg_check_modules(LibFabric libfabric>=1.17)
+pkg_check_modules(LibDPDK libdpdk>=22.11)
 pkg_check_modules(LibIBVerbs libibverbs>=1.14)
 
 #
