+++ src/transport/CMakeLists.txt
@@ -6,9 +6,9 @@ add_subdirectory(pcap)
 add_subdirectory(shm)
 add_subdirectory(stubs)
 
-if (LibFabric_FOUND)
-  add_subdirectory(fabric)
-endif (LibFabric_FOUND)
+if (LibDPDK_FOUND)
+  add_subdirectory(dpdk)
+endif (LibDPDK_FOUND)
 
 if (LibIBVerbs_FOUND)
   add_subdirectory(ofed)
