+++ tools/uspace/dpdk/Connection.h
@@ -2,7 +2,7 @@
 
 #include <utils/Command.h>
 
-namespace tulips::tools::uspace::fabric::connection {
+namespace tulips::tools::uspace::dpdk::connection {
 
 void populate(utils::Commands& cmds);
