+++ tools/uspace/dpdk/State.cpp
@@ -1,6 +1,6 @@
-#include <uspace/fabric/State.h>
+#include <uspace/dpdk/State.h>
 
-namespace tulips::tools::uspace::fabric {
+namespace tulips::tools::uspace::dpdk {
 
 State::State(const bool pcap) : utils::State(), poller(pcap), ids() {}
